Implement HyperellipticCurve.random_element() over finite fields
